stage specific file loading, update sh with gpu stats

+++ src/fp2mol_main.py
@@ -231,7 +231,8 @@ def main(cfg: DictConfig):
                       fast_dev_run=cfg.general.name == 'debug',
                       callbacks=callbacks,
                       log_every_n_steps=50 if name != 'debug' else 1,
-                      logger=loggers)
+                      logger=loggers,
+                      precision="bf16-mixed")
 
     if not cfg.general.test_only:
         trainer.fit(model, datamodule=datamodule, ckpt_path=cfg.general.resume)
